ai-intelligence/requirements.txt
# AI Intelligence UEBA Requirements
# Enterprise Banking SOC - Advanced Analytics

# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
torch==2.0.1
tensorflow==2.13.0
keras==2.13.1

# Anomaly Detection Specialized Libraries
pyod==1.1.0
# isolation-forest and hbos are included in pyod

# Graph Analytics
# py2neo==2021.2.3  # Not compatible with Python 3.11, using neo4j driver instead
networkx==3.1
neo4j==5.11.0

# Time Series Analysis
statsmodels==0.14.0
tslearn==0.6.2
prophet==1.1.4

# API and Web Framework
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.1.1
httpx==0.24.1

# Data Storage and Connectivity
opensearch-py==2.3.1
elasticsearch==8.9.0
redis==4.6.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Configuration and Logging
pyyaml==6.0.1
python-dotenv==1.0.0
structlog==23.1.0
prometheus-client==0.17.1

# Security and Cryptography
cryptography==41.0.3
pycryptodome==3.18.0
jwt==1.3.1

# Data Validation and Serialization
marshmallow==3.20.1
jsonschema==4.19.0

# Utility Libraries
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0
joblib==1.3.1

# Monitoring and Observability
sentry-sdk==1.29.2
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0

# Testing (for development)
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0